Extract graph components and utils to separate package
Upcoming changes to the pipeline graph components will be introducing
new dependencies such as Carbon Charts which I'd like to keep
separate from the rest of the application for now. Extracting the
graph components to their own package makes this easier and
allows consumers to opt-in to the heavier dependencies as needed.

This change is mostly a move/rename with no functional impact.

Also clean up some unwanted entries in the package-lock that were
causing an unnecessary `node_modules/` to be created in the components
package.
