+++ packages/graph/src/components/Graph.scss
