+++ packages/graph/src/components/Graph.stories.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2020 The Tekton Authors
+Copyright 2019-2022 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -12,13 +12,10 @@ limitations under the License.
 */
 
 import React from 'react';
-import { action } from '@storybook/addon-actions';
 
-import PipelineGraph from './PipelineGraph';
+import Graph from './Graph';
 
-import pipeline from './examples/pipeline.json';
-import pipelineRun from './examples/pipelineRun.json';
-import tasks from './examples/tasks.json';
+import graph from './examples/graph.json';
 
 export default {
   parameters: {
@@ -26,15 +23,7 @@ export default {
       default: 'white'
     }
   },
-  title: 'Experimental/Components/Graph/PipelineGraph'
+  title: 'Experimental/Graph/Graph'
 };
 
-export const Base = () => (
-  <PipelineGraph
-    onClickStep={action('onClickStep')}
-    onClickTask={action('onClickTask')}
-    pipeline={pipeline}
-    pipelineRun={pipelineRun}
-    tasks={tasks}
-  />
-);
+export const Base = () => <Graph graph={graph} />;
