+++ packages/graph/src/components/Node.stories.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2021 The Tekton Authors
+Copyright 2019-2022 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -69,7 +69,7 @@ export default {
       default: 'white'
     }
   },
-  title: 'Experimental/Components/Graph/Node'
+  title: 'Experimental/Graph/Node'
 };
 
 export const Start = () => <Node type="Start" width="60" height="60" />;
