+++ packages/graph/src/components/ZoomablePipelineGraph.scss
