+++ packages/graph/src/components/ZoomablePipelineGraph.stories.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2020 The Tekton Authors
+Copyright 2019-2022 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -26,7 +26,7 @@ export default {
       default: 'white'
     }
   },
-  title: 'Experimental/Components/Graph/ZoomablePipelineGraph'
+  title: 'Experimental/Graph/ZoomablePipelineGraph'
 };
 
 export const Base = () => (
