+++ packages/graph/src/components/examples/pipeline.json
