+++ packages/graph/src/index.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2020 The Tekton Authors
+Copyright 2022 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -10,20 +10,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+/* istanbul ignore file */
 
-import React from 'react';
+export { default as buildGraphData } from './buildGraphData';
 
-import Graph from './Graph';
-
-import graph from './examples/graph.json';
-
-export default {
-  parameters: {
-    backgrounds: {
-      default: 'white'
-    }
-  },
-  title: 'Experimental/Components/Graph/Graph'
-};
-
-export const Base = () => <Graph graph={graph} />;
+export { default as Graph } from './components/Graph';
+export { default as PipelineGraph } from './components/PipelineGraph';
